requirements.txt
altair
numpy
pandas
pydeck
streamlit
mysql-connector-python
